--- FLASK_RC1/Pipfile
@@ -1,28 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-
-[packages]
-beautifulsoup4 = "==4.9.0"
-gunicorn = "==20.0.4"
-itsdangerous = "==1.1.0"
-requests = "==2.23.0"
-six = "==1.14.0"
-flask-marshmallow = "==0.12.0"
-psycopg2 = "==2.8.5"
-python-dotenv = "==0.13.0"
-flask-restplus = "==0.13.0"
-Flask = "==1.1.2"
-Jinja2 = "==2.11.2"
-MarkupSafe = "==1.1.1"
-Werkzeug = "==0.16.1"
-Flask-SQLAlchemy = "==2.4.1"
-Flask-Testing = "==0.8.0"
-marshmallow-sqlalchemy = "*"
-flask-compress = "*"
-
-[requires]
-python_version = "3.7"
