--- FLASK_RC1/Procfile
@@ -1,1 +1,0 @@
-web: gunicorn wsgi:app
