--- FLASK_RC1/application/templates/index.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Earth Dashboard</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link href={{ url_for('static', filename='main.css') }} rel="stylesheet" media="screen">
-  </head>
-  <body>
-    <div class="container">
-      <h3>I am working, hear me roar. And also head to /bubblespull to pull
-          data from summary COVID API into the Postgres.
-          <p> and visit /bubblesvisual to view the bubbles visualization.</h3>
-    </div>
-    <
-  </body>
-</html>
