--- FLASK_RC1/config.py
@@ -1,20 +1,0 @@
-from os import environ, path
-from dotenv import load_dotenv
-
-basedir = path.abspath(path.dirname(__file__))
-load_dotenv(path.join(basedir, '.env'))
-
-
-class Config:
-    """Set Flask configuration variables from .env file."""
-
-    # General Flask Config
-    TESTING = environ.get('TESTING')
-    DEBUG = environ.get('DEBUG')
-    FLASK_APP = 'wsgi.py'
-    SECRETKEY = environ.get('SECRETKEY')
-
-    # Database
-    SQLALCHEMY_DATABASE_URI = environ.get('SQLALCHEMY_DATABASE_URI')
-    SQLALCHEMY_ECHO = True
-    SQLALCHEMY_TRACK_MODIFICATIONS = environ.get('SQLALCHEMY_TRACK_MODIFICATIONS')
